I really enjoyed drawing this ship
I have a bad habit of not actually using the software I create, but
whilst perhaps I should have gone to bed, this was a delight to create,
the childlike joy of picking blocks together to form an image, imagining
what I could do with different pieces, and putting silly smiley faces
in as the crew and making them all different colors.

I didn't look at any reference while drawing this, so... it doubtless
contains many inaccuracies.
But that's rather beside the point, isn't it?

Aside from just having fun with it, I observed that the Pick Color tool
is vastly more important than in a raster workflow, since it picks not
just a color, but two colors and a character! And I would really like
the Color Eraser feature. I would sometimes use the wrong blue to match
the sky and then have to Pick Color, change color, and replace the cell,
one by one, whereas the Color Eraser should take care of that.
It's a little awkward selecting colors with Ctrl click, with it
not showing the foreground color (Input doesn't seem to support it), and
with the cognitive dissonance between the background vs foreground
here vs in MS Paint. Other than that, it worked splendidly.