Fix dragging windows
This was broken in "Fix a type checker error" e62064a307fe9dc700d8c7eb591f615b43f7acfb 
where I just added `and self.offset_at_drag_start`.
The problem is that Offset defines __bool__ and are falsy when (0,0),
and the windows, despite starting out in the center, initially have an
offset of (0,0) FROM this center position.