Fix error during undo after redo after undo after resize
When you've finally tracked down an issue and to fix it all you have to
do is point the cursor in the right spot and hit tab to accept the AI's
auto completion, it really makes you wonder about proactive AI coding
assistance, and a future with AGI. Or even without any AGI or proactive
AI, just having a gutter view that shows how surprised a language model
is at a given line of code could really work wonders on a problem like
this. It might feel _judgy_, but it could solve problems.