Automatically recover from backup file
- Recover from backup file if it exists for the document when loading
  a document via the CLI or TUI, or if it exists for Untitled when
  creating a new document.
- Reorganize CLI arguments handling into passive and active,
  so that the active recovery step can use the passive* backup folder
  option. (*Directories are created if they don't exist.)

Working on automatically recovering the document,
I'm realizing this is more complicated than I thought.
Consider two instances of the app (A and B) editing the same file:

- Make changes in A
- Open same file in B
- A auto-saves after timer
- Close B
- B deletes auto-save when closing
- A doesn't re-create the auto-save unless further changes are made to
  the document, because of the undo tracking, which is basically an
  optimization to avoid writing to the disk all day while idle.

Also, there could be a race condition where it saves over the auto save
before recovering it — potentially — like if the auto save interval
was implemented such that it made an immediate call at the start, or
if some other app code tries to auto save initially.
Or, consider if you:

- Make changes in A
- Open same file in B
- Make changes in B, maybe before it detects the backup file if the disk
  is very slow for some reason, or if you just ignore the prompt.
  (None of the prompts are modal, currently.)

To address these issues, I could add a flag that says whether the
auto save was written, and thus owned by, this session, and make it
not write to a backup file that it didn't create in the current session.
But it still ideally would auto save, just to a different file,
so I could instead have a map of file paths to bools saying whether it's
owned by the app or not and then have a number incremented in the
filename (maybe like `.ans~` and then `.ans~SESSION2~`, including
"SESSION" so it's clear it's not time-based), adding 1 until it's a file
that can be owned by the session when auto saving.
But if an earlier-numbered session closes, you wouldn't want it to leave
a file behind. Either it should delete/rename its backup file to the
earlier number and start saving to that, or it should keep the existing
higher number. I guess if looks at the map first, before checking the
file system, it shouldn't be a problem. It should keep using the higher
number. Would this be a little overkill? Eh, maybe.