Cache Style objects for each cell in the document
This is not worth the complexity. It would be better to just change the document model to replace fg and bg with a 2D array of Style objects, so there's still a single source of truth.

This touches a LOT of code, and cache invalidation is classically known as a hard problem in programming.