Make backspace delete to left in free typing mode
I reworked this so it edits the image directly, creating undo states
immediately when you type instead of when the textbox is later melded.
Now it can edit a different cell without complicating the code, as the
edits are done directly within the cursor movement code instead of
later when the 1x1 textbox is melded into the canvas.