Fix regression where CharInput defaults to focused
Originally the File menu was getting the default focus, though this was
not useful because the menus are not keyboard accessible yet, but when I
made the menu items not accept focus, the auto focus behavior became to
focus the CharInput, causing `pytest -k test_paint_app` to fail.
No focus makes more sense than the File menu item or the CharInput.